[Raghav Bobal] made some changes in mobile class

+++ C0780996_W2020_MID_TERM_MAD3004/C0780996_W2020_MID_TERM_MAD3004/Bills/Mobile.swift
@@ -13,12 +13,12 @@ public class Mobile: Bill
     private var mobileManufacturerName: String
     private var planName: String
     private var mobileMumber: String
-    private var mobileDataUsed: Float
-    private var callMinutesUsed: Float
+    private var mobileDataUsed: Double
+    private var callMinutesUsed: Double
     var type: BillType = BillType.Mobile
 
     
-     init(billId: Int, billDate: Date?, billType: BillType,customerId: Int,mobileManufacturerName: String, planName: String, mobileMumber: String, mobileDataUsed: Float, callMinutesUsed: Float)
+     init(billId: Int, billDate: Date?, billType: BillType,customerId: Int,mobileManufacturerName: String, planName: String, mobileMumber: String, mobileDataUsed: Double, callMinutesUsed: Double)
     {
            self.mobileManufacturerName = mobileManufacturerName
            self.planName = planName
